2020.7.8

+++ methods/baselinefinetune.py
@@ -42,10 +42,10 @@ class BaselineFinetune(MetaTemplate):
             for i in range(0, support_size , batch_size):
                 set_optimizer.zero_grad()
                 selected_id = torch.from_numpy( rand_id[i: min(i+batch_size, support_size) ]).cuda()
-                z_batch = z_support[selected_id]
-                y_batch = y_support[selected_id] 
+                z_batch = z_support[selected_id.long()]
+                y_batch = y_support[selected_id.long()] 
                 scores = linear_clf(z_batch)
-                loss = loss_function(scores,y_batch)
+                loss = loss_function(scores,y_batch.long())
                 loss.backward()
                 set_optimizer.step()
         scores = linear_clf(z_query)
